modularised forward and inverse gas cycles
